Prepare for R 2.14
According to John Fox, we'd better make a few small changes to
work under the upcoming R 2.14:
- use .onAttach() instead of .First.lib()
- use a namespace

This allows us to drop the custom gettext_() wrapper, as the
namespacing passes the domain automatically.

We don't import R2HTML as it's better to load it only when needed,
i.e. when copying a table: it's not always used.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@35 edb9625f-4e0d-4859-8d74-9fd3b1da38cb