Factiva: Use shorter article IDs
The original ID contains too many useless characters. We can safely
remove the redundant 0, and almost as safely replace the long serial
number with the last two chars. A collision can still happen if two
articles from the same origin and the same day have the same last
two chars, but the general scheme seems to be that these last two chars
change first, and the beginning of the serial number identifies the larger
series, which means it would not happen.

A better solution would be to attribute unique letters from
FactivaSource once we know all the IDs, but this is not possible with
XMLSource currently.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@103 edb9625f-4e0d-4859-8d74-9fd3b1da38cb