Use wait cursor for long operations
Rcmdr has no way to report activity, and users can wonder what's
going on. Only do this for potentially long operations, to
avoid disturbing casual operations. Rcmdr could do this
automatically for all operations in justDoIt() and doItAndPrint(),
but 1) it would mean the cursor would blink when many steps are run,
and 2) in my tests under Linux, the cursor is not updated at all,
probably because of a bug in the Tcl/Tk event loop (same problem as
with current command not being printed to the log until after the
operation completes).

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@117 edb9625f-4e0d-4859-8d74-9fd3b1da38cb