Rework clustering
Use Chi-squared distance instead of pure euclidean one: this is
more robust WRT the size of texts, and is more consistent with CA.
Improve plot by using translated legends, a more explicit ylab,
and hiding leaf labels when there are too many texts. Rename "class"
to "cluster" as it's the common word in English. Fix a bug when
removeSparseTerms() creates docs with no term.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@165 edb9625f-4e0d-4859-8d74-9fd3b1da38cb