Work around DirSource() using UTF-8 as encoding
DirSource() has a encoding argument defaulting to UTF-8, which is
passed to readLines() in getElem.DirSource(). This affects the marked
encoding of the internal string (not the encoding used when reading
the file). On non UTF-8 systems (Windows), this means that .txt files
that were correctly read using the locale encoding are then incorrectly
printed because the declared encoding is wrong.

This change does not affect Linux, even though the encoding for the
documents is now unknown.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@234 edb9625f-4e0d-4859-8d74-9fd3b1da38cb