Allow more tables to be exported
Some tables were forgotten when the feature was added, or were
not supported (lists). Add a special case for lists, since the default
HTML.list() does not print the element names, does not explicitly say when
a vector is empty, and redirects align="left" to cat(); for arrays, use
HTML.list() too since the default method does not work in all cases.

New objects need to be removed when doing cleanup.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@243 edb9625f-4e0d-4859-8d74-9fd3b1da38cb