Convert variable names to lower case for plot titles
In most cases, the variable name is a common word that integrates
better in the sentence if in lowercase. This is especially true
for standard variables like Origin and Date. It is not so nice for
topics and geographic areas, but these boolean variables do not give
nice plots as-is anyway.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@266 edb9625f-4e0d-4859-8d74-9fd3b1da38cb