Add terms dictionary with stemming results and fix import details
This prompted a few fixes: we need to handle thin spaces and thin
non-breaking spaces expicitly or they get concatenated with words
(at least in the dictionary). We also have to pass language explicitly
to stemDocument(), else the document's language is used, which may not
be set correctly. Finally, words of length 2 are not kept by default,
while this may be useful, so pass the option to DocumentTermMatrix().

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@311 edb9625f-4e0d-4859-8d74-9fd3b1da38cb