Work around a bug in Rcmdr on import when setting 'source' tag
Rcmdr has a bug when splitting commands with more than one pair of
quote and a long quoted string: the quoted string can get split in
some rare cases (narrow window), and trigger an "unbalanced quote"
error that halts the import. Until the fix is in a released Rcmdr
version, better set the tag invisibly: this means the source file
will be empty if the import is run again from the code.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@363 edb9625f-4e0d-4859-8d74-9fd3b1da38cb