Add setLastTable() to remember last table and set its title
This makes the code cleaner, and allows removing R CMD check warnings
due to the use of "<<-". Now we use doItAndPrint() to set the title,
and print the table after setting it, so that it is printed in the
console.

Also remove explicit print(..., digits=2) since we set this via
options(digits) now.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@390 edb9625f-4e0d-4859-8d74-9fd3b1da38cb