tm.plugin.factiva: Extract company, industry, IPC and IPD
Based on a patch by Grigorij Ljubownikow. In readFactivaHTML(),
adapt the code to handle three-letter field codes. For XML,
we add to IPD every information from fieldDesc that does not
fit into IPC: this seems to be what happens with HTML, and it
avoid discarding the many different fields that exist (cf.
http://www.factiva.com/cp/downloads/select/nov2004/xml/metadatapt.desctpc.descfield.htm).

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@444 edb9625f-4e0d-4859-8d74-9fd3b1da38cb