Move package methods to Depends
On Windows 7, using R 3.1.0, having it in Imports is not enough
when using the shortcut provided by the R.TeMiS installer.
Manual edition of stemming fails with a "could not find function
'loadMethod'" error. Calling library(methods) fixes it. This is
because the shortcut uses Rscript, which doesn't load methods
by default. The bug does not appear with the default R console,
which attaches methods by default.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@562 edb9625f-4e0d-4859-8d74-9fd3b1da38cb